1. Add "get" function to leaky bucket that returns the current values 2. fix bug where if the leaky bucket is "off" (MAX_LONG) the result would allocate too much memory for the receive buffer. 3. Add "gettrafficshaping" CLI command 4. Add "settrafficshaping send|receive disable" command to stop shaping

+++ src/net.h
@@ -51,7 +51,7 @@ static const unsigned int MAX_INV_SZ = 50000;
 /** The maximum number of new addresses to accumulate before announcing. */
 static const unsigned int MAX_ADDR_TO_SEND = 1000;
 /** The maximum # of bytes to receive at once */
-static const int MAX_RECV_CHUNK = 256*1024;
+static const int64_t MAX_RECV_CHUNK = 256*1024;
 
 /** -listen default */
 static const bool DEFAULT_LISTEN = true;
